# proyecto/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TaskBoard - Trello-like project management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImage?: string | null;
  isAuthenticated: boolean;
}

export interface BoardSummary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  memberCount: number;
  cardCount: number;
  createdAt: string;
  updatedAt: string;
  role: string;
}

export interface Card {
  id: number;
  boardId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  position: number;
  /** @nullable */
  assigneeId?: string | null;
  /** @nullable */
  assigneeUsername?: string | null;
  /** @nullable */
  assigneeImage?: string | null;
  /** @nullable */
  dueDate?: string | null;
  attachmentCount: number;
  linkCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface BoardMember {
  id: number;
  boardId: number;
  userId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImage?: string | null;
  role: string;
  joinedAt: string;
}

export interface Board {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  cards: Card[];
  members: BoardMember[];
}

export interface BoardInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  color?: string;
}

export interface BoardUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
}

export interface BoardStats {
  boardId: number;
  todoCount: number;
  inProgressCount: number;
  doneCount: number;
  totalCards: number;
  memberCount: number;
}

export interface ActivityEvent {
  id: number;
  boardId: number;
  /** @nullable */
  cardId?: number | null;
  /** @nullable */
  cardTitle?: string | null;
  userId: string;
  userDisplayName: string;
  action: string;
  /** @nullable */
  detail?: string | null;
  createdAt: string;
}

export interface InvitationInput {
  /** @minLength 1 */
  username: string;
}

export interface Invitation {
  id: number;
  boardId: number;
  boardName: string;
  /** @nullable */
  invitedByUsername?: string | null;
  status: string;
  createdAt: string;
}

export interface CardAttachment {
  id: number;
  cardId: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  url: string;
  createdAt: string;
}

export interface CardLink {
  id: number;
  cardId: number;
  url: string;
  /** @nullable */
  title?: string | null;
  createdAt: string;
}

export interface CardDetail {
  id: number;
  boardId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  position: number;
  /** @nullable */
  assigneeId?: string | null;
  /** @nullable */
  assigneeUsername?: string | null;
  /** @nullable */
  assigneeImage?: string | null;
  /** @nullable */
  dueDate?: string | null;
  attachments: CardAttachment[];
  links: CardLink[];
  createdAt: string;
  updatedAt: string;
}

export type CardInputStatus = typeof CardInputStatus[keyof typeof CardInputStatus];


export const CardInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface CardInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: CardInputStatus;
  assigneeId?: string;
  dueDate?: string;
}

export type CardUpdateStatus = typeof CardUpdateStatus[keyof typeof CardUpdateStatus];


export const CardUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface CardUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: CardUpdateStatus;
  position?: number;
  /** @nullable */
  assigneeId?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface CardLinkInput {
  url: string;
  title?: string;
}